--- frontend/src/components/PageTitle.tsx
@@ -1,8 +1,0 @@
-function PageTitle()
-{
-    return (
-        <h1 id="title">COP 4331 MERN Stack Demo</h1>
-    );
-};
-
-export default PageTitle;
