+++ frontend/src/pages/VerifyPage.tsx
@@ -1,10 +1,8 @@
-import PageTitle from "../components/PageTitle.tsx";
 import EmailVerify from "../components/EmailVerify.tsx";
 
 const VerifyPage = () => {
     return (
         <div>
-            <PageTitle />
             <EmailVerify />
         </div>
     );
